+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Big things planned! More unit tests! Less ORM quirks! More awesomeness!
 * **services:** add ErrorHandler service
 * **router:** allow ErrorHandler to attempt to handle errors before they propogate
 * **router:** allow ErrorHandler to attempt to handle routes that did not send a response
+* **crud-controller:** add policies returned from getDestroyPolicies to the destroy route
 
 ### Bug Fixes